feat(dependency): increase minimum version of dep
Increase the minimum version for setuptools-scm.

This helps consumers of hatch-vcs to test their package with the minimum
suported versions of dependencies.
Earlier versions of setuptools-scm did depend on setuptools without
a version specifier. Using a tool that allows you to install the
earliest supported version of every packages ends up installing
setuptools 0.x which would fail of course.
Starting with setuptools-scm 8.2.0, the minimum version is 60 something.